+++ framework/scenario/src/scenario/tx_to_step.rs
@@ -5,6 +5,7 @@ mod tx_to_step_deploy;
 mod tx_to_step_query;
 mod tx_to_step_trait;
 mod tx_to_step_transfer;
+mod tx_to_step_upgrade;
 
 pub use step_annotation::*;
 pub use step_wrapper::{StepWithResponse, StepWrapper};
