+++ framework/scenario/src/scenario/tx_to_step/tx_to_step_call.rs
@@ -1,13 +1,9 @@
 use multiversx_sc::types::{
-    Code, FunctionCall, NotPayable, RHListExec, Tx, TxEnv, TxEnvWithTxHash, TxFromSpecified, TxGas,
-    TxPayment, TxToSpecified, UpgradeCall,
+    FunctionCall, RHListExec, Tx, TxEnv, TxEnvWithTxHash, TxFromSpecified, TxGas, TxPayment,
+    TxToSpecified,
 };
 
-use crate::{
-    imports::MxscPath,
-    scenario_model::{ScCallStep, TxESDT, TxExpect, TxResponse},
-    ScenarioEnvExec,
-};
+use crate::scenario_model::{ScCallStep, TxESDT, TxExpect, TxResponse};
 
 use super::{address_annotated, gas_annotated, StepWrapper, TxToStep};
 
@@ -77,56 +73,6 @@ where
             .iter()
             .map(TxESDT::from)
             .collect();
-    }
-
-    step
-}
-
-impl<'w, From, To, RH> TxToStep<ScenarioEnvExec<'w>, RH>
-    for Tx<
-        ScenarioEnvExec<'w>,
-        From,
-        To,
-        NotPayable,
-        (),
-        UpgradeCall<ScenarioEnvExec<'w>, Code<MxscPath<'w>>>,
-        RH,
-    >
-where
-    From: TxFromSpecified<ScenarioEnvExec<'w>>,
-    To: TxToSpecified<ScenarioEnvExec<'w>>,
-    RH: RHListExec<TxResponse, ScenarioEnvExec<'w>>,
-{
-    type Step = ScCallStep;
-
-    fn tx_to_step(self) -> StepWrapper<ScenarioEnvExec<'w>, Self::Step, RH> {
-        let mut step = tx_to_sc_call_upgrade_step(&self.env, self.from, self.to, self.data);
-        step.expect = Some(self.result_handler.list_tx_expect());
-
-        StepWrapper {
-            env: self.env,
-            step,
-            result_handler: self.result_handler,
-        }
-    }
-}
-
-pub fn tx_to_sc_call_upgrade_step<'a, 'w: 'a, From, To>(
-    env: &'a ScenarioEnvExec<'w>,
-    from: From,
-    to: To,
-    data: UpgradeCall<ScenarioEnvExec<'w>, Code<MxscPath>>,
-) -> ScCallStep
-where
-    From: TxFromSpecified<ScenarioEnvExec<'w>>,
-    To: TxToSpecified<ScenarioEnvExec<'w>>,
-{
-    let mut step = ScCallStep::new()
-        .from(address_annotated(env, &from))
-        .to(address_annotated(env, &to))
-        .function("upgrade");
-    for arg in data.arg_buffer.iter_buffers() {
-        step.tx.arguments.push(arg.to_vec().into());
     }
 
     step
